+++ glfw/wl_platform.h
@@ -233,7 +233,7 @@ typedef struct _GLFWwindowWayland
         struct {
             int width, height;
             bool focused, needs_shadow;
-            float fscale;
+            double fscale;
             WaylandWindowState toplevel_states;
         } for_window_state;
 
@@ -423,7 +423,7 @@ void _glfwAddOutputWayland(uint32_t name, uint32_t version);
 void _glfwWaylandAfterBufferSwap(_GLFWwindow *window);
 void _glfwSetupWaylandDataDevice(void);
 void _glfwSetupWaylandPrimarySelectionDevice(void);
-float _glfwWaylandWindowScale(_GLFWwindow*);
+double _glfwWaylandWindowScale(_GLFWwindow*);
 int _glfwWaylandIntegerWindowScale(_GLFWwindow*);
 void animateCursorImage(id_type timer_id, void *data);
 struct wl_cursor* _glfwLoadCursor(GLFWCursorShape, struct wl_cursor_theme*);
